Program structure
Worked on url structure of the project. I will take a second pass over this code tomorrow to make sure I didn't miss anything.

TODO:
themes
pages

Update note:
It looks like Dash just updated to 2.0.0 approximately 5 days ago. This is currently breaking dash-bootstrap-components as they have not released an update that is fully compatible with v2.0.0. I think the smartest option would be to revert back to the latest stable Dash v1.* build.

I will try to do some digging and determine when an update will be pushed for dash-bootstrap-components. If an update is happening soon, we will just update dash-bootstrap-components.